% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_extract.R
\name{as_tibble.kagi_results}
\alias{as_tibble.kagi_results}
\alias{as_tibble.kagi_search_results}
\title{Convert Kagi results to a tibble}
\usage{
as_tibble.kagi_results(x, ...)

as_tibble.kagi_search_results(x, ...)
}
\arguments{
\item{x}{A kagi_results object}

\item{...}{Additional parameters passed to methods}
}
\value{
A tibble representation of the results
}
\description{
Convert Kagi results to a tibble
}
