% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_extract.R
\name{kagi_hits}
\alias{kagi_hits}
\alias{kagi_hits.kagi_search_results}
\title{Extract search hits from Kagi search results}
\usage{
kagi_hits(x, ...)

\method{kagi_hits}{kagi_search_results}(x, ...)
}
\arguments{
\item{x}{A kagi_search_results object}

\item{...}{Additional parameters (not used)}
}
\value{
A tibble of search hits
}
\description{
Extract search hits from Kagi search results
}
