% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_extract.R
\name{kagi_meta}
\alias{kagi_meta}
\alias{kagi_meta.kagi_search_results}
\title{Extract metadata from Kagi results}
\usage{
kagi_meta(x, ...)

\method{kagi_meta}{kagi_search_results}(x, ...)
}
\arguments{
\item{x}{A kagi_results object}

\item{...}{Additional parameters passed to methods}
}
\value{
A list of metadata
}
\description{
Extract metadata from Kagi results
}
