% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_perform.R, R/generics.R,
%   R/search_perform.R, R/summarize_perform.R
\name{kagi_perform.kagi_enrich}
\alias{kagi_perform.kagi_enrich}
\alias{kagi_perform}
\alias{kagi_perform.kagi_search}
\alias{kagi_perform.kagi_summarize}
\title{Execute a Kagi API request}
\usage{
\method{kagi_perform}{kagi_enrich}(x, path = NULL, ...)

kagi_perform(x, path, ...)

\method{kagi_perform}{kagi_search}(x, path = NULL, ...)

\method{kagi_perform}{kagi_summarize}(x, path = NULL, ...)
}
\arguments{
\item{x}{The request object to execute}

\item{path}{Optional path to a directory where the raw JSON should be saved}

\item{...}{Additional parameters passed to methods}
}
\value{
A results object corresponding to the request type
}
\description{
Dispatches to endpoint-specific methods based on the class of `x`.
See the method documentation for details on request shape and returned
result classes.
}
\section{Methods (by class)}{
\itemize{
\item \code{kagi_perform(kagi_enrich)}: Execute an Enrich request

\item \code{kagi_perform(kagi_search)}: Execute a Search request

\item \code{kagi_perform(kagi_summarize)}: Execute a Summarize request

}}
