% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_extract.R
\name{kagi_related}
\alias{kagi_related}
\alias{kagi_related.kagi_search_results}
\title{Extract related searches from Kagi results}
\usage{
kagi_related(x, ...)

\method{kagi_related}{kagi_search_results}(x, ...)
}
\arguments{
\item{x}{A kagi_results object}

\item{...}{Additional parameters passed to methods}
}
\value{
A data structure containing the related items (type depends on method)
}
\description{
Extract related searches from Kagi results
}
