% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_query.R
\name{summarize_query}
\alias{summarize_query}
\title{Create a new Kagi summarize request}
\usage{
summarize_query(
  url = NULL,
  text = NULL,
  engine = NULL,
  summary_type = NULL,
  target_language = NULL,
  cache = TRUE
)
}
\arguments{
\item{url}{Optional character scalar. URL to be summarized. Mutually exclusive
with \code{text}.}

\item{text}{Optional character scalar. Raw text to be summarized. Mutually
exclusive with \code{url}.}

\item{engine}{Character scalar. Summarizer engine (options: \code{"cecil"},
\code{"agnes"}, \code{"muriel"}, \code{"daphne"}). Default: \code{"cecil"}.}

\item{summary_type}{Character scalar. Type of summary requested (options:
\code{"summary"}, \code{"takeaway"}). Default: \code{"summary"}.}

\item{target_language}{Character scalar. Target language ISO code. Supported
codes: \code{"EN"}, \code{"BG"}, \code{"CS"}, \code{"DA"}, \code{"DE"}, \code{"EL"}, \code{"ES"}, \code{"ET"},
\code{"FI"}, \code{"FR"}, \code{"HU"}, \code{"ID"}, \code{"IT"}, \code{"JA"}, \code{"KO"}, \code{"LT"}, \code{"LV"},
\code{"NB"}, \code{"NL"}, \code{"PL"}, \code{"PT"}, \code{"RO"}, \code{"RU"}, \code{"SK"}, \code{"SL"}, \code{"SV"},
\code{"TR"}, \code{"UK"}, \code{"ZH"}, \code{"ZH-HANT"}. Default: \code{"EN"}.}

\item{cache}{Logical. Whether to allow API-side caching.}
}
\value{
A named list of \code{kagi_summarize_query} objects to be passed to
\code{\link[=kagi_request]{kagi_request()}}.
}
\description{
Construct a typed S3 object of class \code{kagi_summarize} that describes a
Universal Summarizer request. Use \code{\link[=kagi_request]{kagi_request()}} to execute the request
and obtain the json replies.
}
\examples{
\dontrun{
req <- summarize_query(text = "Lorem ipsum")
req
}

}
